Match a lot of calc.c, add labels to Shift-JIS font (#3)
* Match a lot of calc.c, add labels to Shift-JIS font

* Review

* Improve mtxf format

+++ include/main_segment_functions.h
@@ -888,7 +888,7 @@ s16 angleF2S(f32 arg0);
 // void func_8007CFB4();
 void randomize00(void);
 // void randomseed();
-u16 irandom(void);
+s32 irandom(void);
 // void random();
 
 // void func_8007D260();
